fix protected routes
